backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
pyjwt==2.1.0
werkzeug==2.0.1